refactor: rewrite NotFoundPage component with updated metadata and OpenGraph support
